globalize mapLayers
