--- Minge2022Summer/src/Scenes/Objects/Stair.hpp
@@ -1,3 +1,0 @@
-﻿#pragma once
-# include "Object.hpp"
-
